requirements.txt
discord.py
python-dotenv

#시간
pytz